requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
argon2-cffi=20.1.0=py37h27cfd23_1
async_generator=1.10=py37h28b3542_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
blosc=1.21.0=h8c45485_0
brotli=1.0.9=he6710b0_2
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.4.13=h06a4308_1
certifi=2020.12.5=py37h06a4308_0
cffi=1.14.5=py37h261ae71_0
charls=2.2.0=h2531618_0
cloudpickle=1.6.0=py_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2021.4.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
decorator=5.0.6=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py37_0
expat=2.3.0=h2531618_2
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
fsspec=0.9.0=pyhd3eb1b0_0
giflib=5.1.4=h14c3975_1
glib=2.68.1=h36276a3_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
imagecodecs=2021.3.31=py37h581e88b_0
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=3.10.0=py37h06a4308_0
importlib_metadata=3.10.0=hd3eb1b0_0
iniconfig=1.1.1=pyhd3eb1b0_0
intel-openmp=2021.2.0=h06a4308_610
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.22.0=py37hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3eb1b0_1
jedi=0.17.0=py37_0
jinja2=2.11.3=pyhd3eb1b0_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py_2
jupyter=1.0.0=py37_7
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.7.1=py37h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
jxrlib=1.1=h7b6447c_2
kiwisolver=1.3.1=py37h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
lerc=2.2.1=h2531618_0
libaec=1.0.4=he6710b0_1
libdeflate=1.7=h27cfd23_5
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libwebp=1.0.1=h8e7db2f_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
libzopfli=1.0.3=he6710b0_0
locket=0.2.1=py37h06a4308_1
lz4-c=1.9.3=h2531618_0
markupsafe=1.1.1=py37h14c3975_1
matplotlib-base=3.3.4=py37h62a2d02_0
mistune=0.8.4=py37h14c3975_1001
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.1.1=py37h0573a6f_0
more-itertools=8.7.0=pyhd3eb1b0_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py37_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.5=py_0
ninja=1.10.2=hff7bd54_1
notebook=6.3.0=py37h06a4308_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
olefile=0.46=py37_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1k=h27cfd23_0
packaging=20.9=pyhd3eb1b0_0
pandoc=2.12=h06a4308_0
pandocfilters=1.4.3=py37h06a4308_1
parso=0.8.2=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py37he98fc37_0
pip=21.0.1=py37h06a4308_0
pluggy=0.13.1=py37h06a4308_0
prometheus_client=0.10.1=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
prompt_toolkit=3.0.17=hd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py=1.10.0=pyhd3eb1b0_0
pycparser=2.20=py_2
pygments=2.8.1=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.17.3=py37h7b6447c_0
pytest=6.2.2=py37h06a4308_2
python=3.7.10=hdb3f193_0
python-dateutil=2.8.1=pyhd3eb1b0_0
pytorch=1.3.1=cuda100py37h53c1284_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.4.1=py37h27cfd23_1
pyzmq=20.0.0=py37h2531618_1
qt=5.9.7=h5867ecd_1
qtconsole=5.0.3=pyhd3eb1b0_0
qtpy=1.9.0=py_0
readline=8.1=h27cfd23_0
scikit-image=0.18.1=py37ha9443f7_0
scikit-learn=0.24.1=py37ha9443f7_0
scipy=1.6.2=py37h91f5cce_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py37h06a4308_0
snappy=1.1.8=he6710b0_0
sqlite=3.35.4=hdfb4753_0
terminado=0.9.4=py37h06a4308_0
testpath=0.4.4=pyhd3eb1b0_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2021.4.8=pyhd3eb1b0_2
tk=8.6.10=hbc83047_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.11.1=pyhd3eb1b0_0
torchvision=0.4.2=cuda100py37hecfc37a_0
tornado=6.1=py37h27cfd23_0
traitlets=5.0.5=pyhd3eb1b0_0
typing_extensions=3.7.4.3=pyha847dfd_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py37_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zfp=0.5.5=h2531618_6
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
